+++ tests/test_layouts.py
@@ -1,6 +1,7 @@
 import unittest
 from math import hypot
 from igraph import Graph, Layout, BoundingBox, InternalError
+from igraph import umap_compute_weights
 
 
 class LayoutTests(unittest.TestCase):
@@ -317,6 +318,17 @@ class LayoutAlgorithmTests(unittest.TestCase):
         # Same but inputting the coordinates
         lo_adj = g.layout_umap(dist=dist, epochs=1, seed=lo.coords)
         self.assertTrue(isinstance(lo_adj, Layout))
+
+    def testUMAPComputeWeights(self):
+        edges = [0, 1, 0, 2, 1, 2, 1, 3, 2, 3, 2, 0]
+        edges = list(zip(edges[::2], edges[1::2]))
+        dist = [1, 1.5, 1.8, 2.0, 3.4, 0.5]
+        # NOTE: you need a directed graph to make sense of the symmetryzation
+        g = Graph(edges, directed=True)
+        weights = umap_compute_weights(g, dist)
+        self.assertEqual(
+            weights,
+            [1.0, 1.0, 1.0, 1.1253517471925912e-07, 6.14421235332821e-06, 0.0])
 
     def testLGL(self):
         g = Graph.Barabasi(100)